server/requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
openslide-python==1.2.0
Pillow==9.4.0
Werkzeug==2.2.3
zipp==3.15.0
